metrics_collector/metrics_collector/requirements.txt
linux-metrics==0.1.4
kafka-python==1.4.7
-e git+https://github.com/serverdensity/python-daemon.git@843cb59#egg=daemon